pyproject.toml
[project]
name = "data-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dataframe-image>=0.2.7",
    "datasets>=3.3.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "google-generativeai>=0.8.4",
    "matplotlib>=3.10.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "ruff>=0.9.9",
    "selenium>=4.29.0",
    "streamlit>=1.43.2",
    "termcolor>=2.5.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.49.0",
]
